criando componentes e página notfound
