--- teste/Components/header.jsx
@@ -1,15 +1,0 @@
-export default function Header(){
-
-    return(
-
-        <header>
-            <h1>Minha aplicação </h1>
-            <nav>
-                <a href="/">Início</a>
-                <a href="/about">Sobre</a>
-            </nav>
-        </header>
-
-    )
-
-}
