--- teste/app/About/page.tsx
@@ -1,11 +1,0 @@
-export default function About(){
-
-    return(
-
-        <main>
-            <h1>Essa é a página sobre</h1>
-        </main>
-
-    )
-
-}
