--- teste/app/page.tsx
@@ -1,12 +1,0 @@
-export default function Home(){
-
-  return(
-
-    <main>
-        <h1>Bem vindo ao meu site</h1>
-        <p>Explorando o app router</p>
-    </main>
-    
-  )
-
-}
