+++ teste/postcss.config.mjs
@@ -1,8 +1,5 @@
-/** @type {import('postcss-load-config').Config} */
 const config = {
-  plugins: {
-    tailwindcss: {},
-  },
+  plugins: ["@tailwindcss/postcss"],
 };
 
 export default config;
