+++ teste/src/app/contatos/page.jsx
@@ -1,0 +1,11 @@
+export default function Contato(){
+
+    return(
+
+        <div>
+            <h2>Página de contato</h2>
+        </div>
+
+    )
+
+}
