+++ teste/src/app/layout.tsx
@@ -1,0 +1,17 @@
+import "./globals.css";
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body
+        className={`antialiased`}
+      >
+        {children}
+      </body>
+    </html>
+  );
+}
