+++ teste/src/app/page.tsx
@@ -1,0 +1,12 @@
+import { Header } from "@/components/header"
+
+export default function Home(){
+    return(   
+        <div>
+          <Header/>
+          <h2>Olá mundo</h2>
+        </div>
+
+    )
+
+}
