+++ teste/src/components/header/index.tsx
@@ -1,0 +1,8 @@
+export function Header() {
+    return (
+        <header>
+            <h2>Header do projeto</h2>
+        </header>
+    )
+
+}
